mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/server"]

[project]
name = "ug-office-mcp"
version = "0.1.0"
description = "FastMCP bridge to UG Office sportsbook back-office API"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.0.0b2",
    "httpx~=0.27",
    "playwright~=1.40",
    "pydantic~=2.0",
    "python-dotenv~=1.0",
    "returns>=0.23",
    "cytoolz>=1.0",
    "python-socketio[asyncio_client]>=5.0",
]

[project.optional-dependencies]
discovery = []
dev = ["pytest", "pytest-asyncio", "respx", "ruff"]

[project.scripts]
ug-office-mcp = "server.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I"]

[tool.ruff.lint.isort]
known-first-party = ["server", "discovery"]
